implemented Eccentricity calculations with Breeze SparseVectors instead of Scala maps

+++ scala/plongeur-spark/src/main/scala/org/tmoerman/plongeur/tda/Model.scala
@@ -85,7 +85,7 @@ object Model {
 
     val self = this
 
-    lazy val N = dataPoints.count
+    lazy val N = dataPoints.count.toInt
 
     lazy val dim = dataPoints.first.features.size
